# docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  internal_http.BalancePoint:
    properties:
      balance:
        example: 4210.88
        type: number
      date:
        example: "2025-01-31"
        type: string
    type: object
  internal_http.CategorySlice:
    properties:
      amount:
        example: 612.3
        type: number
      category:
        example: Groceries
        type: string
      share:
        description: Share is the fraction of total window spending, 0..1.
        example: 0.18
        type: number
      transaction_count:
        example: 27
        type: integer
    type: object
...
  internal_http.EssentialSplitPoint:
    properties:
      discretionary:
        example: 830.55
        type: number
      essential:
        example: 1580
        type: number
      month:
        example: 2025-01
        type: string
    type: object
...
  internal_http.InvalidParam:
    properties:
      name:
        type: string
      reason:
        type: string
    type: object
  internal_http.MonthCategoryPoint:
    properties:
      amount:
        example: 204.1
        type: number
      category:
        example: Groceries
        type: string
      month:
        example: 2025-01
        type: string
    type: object
  internal_http.MonthlyIncomeSpendingPoint:
    properties:
      income:
        example: 3200
        type: number
      month:
        example: 2025-01
        type: string
      net:
        description: |-
          Net is income minus spending: positive means the month added to
          savings.
        example: 789.45
        type: number
      spending:
        example: 2410.55
        type: number
    type: object
  internal_http.Period:
    properties:
      from:
        example: 2025-01
        type: string
      to:
        example: 2025-03
        type: string
    type: object
  internal_http.PresentCharts:
    properties:
      balance_over_time:
        items:
          $ref: '#/definitions/internal_http.BalancePoint'
        type: array
      category_breakdown:
        items:
          $ref: '#/definitions/internal_http.CategorySlice'
        type: array
      essential_vs_discretionary:
        items:
          $ref: '#/definitions/internal_http.EssentialSplitPoint'
        type: array
      monthly_by_category:
        items:
          $ref: '#/definitions/internal_http.MonthCategoryPoint'
        type: array
      monthly_income_vs_spending:
        items:
          $ref: '#/definitions/internal_http.MonthlyIncomeSpendingPoint'
        type: array
    type: object
  internal_http.Problem:
    properties:
      detail:
        description: |-
          Detail explains this specific occurrence. It never contains
          credentials or internal stack traces.
        type: string
      instance:
        description: Instance identifies this occurrence — here, the request path.
        type: string
      invalid_params:
        description: |-
          InvalidParams is an extension member listing per-field validation
          failures.
        items:
          $ref: '#/definitions/internal_http.InvalidParam'
        type: array
      request_id:
        description: |-
          RequestID is an extension member that ties a response to its log
          lines.
        type: string
      status:
        description: Status repeats the HTTP status code, so the body stands alone.
        type: integer
      title:
        description: Title is a short, human-readable summary, stable for a given
          Type.
        type: string
      type:
        description: |-
          Type is a URI reference identifying the problem kind. "about:blank"
          means the status code alone describes it, which RFC 7807 defines as
          the default.
        type: string
    type: object
...
  internal_http.VisualizationResponse:
    properties:
      currency:
        example: EUR
        type: string
      forecast:
        description: |-
          Forecast is served verbatim from the analysis's chart_forecast
          column: it is the model's projection, not something that can be
          recomputed from recorded transactions. It is null when the model
          produced no forecast.
        type: object
      period:
        $ref: '#/definitions/internal_http.Period'
      present:
        $ref: '#/definitions/internal_http.PresentCharts'
      summary:
        $ref: '#/definitions/internal_http.VisualizationSummary'
    type: object
  internal_http.VisualizationSummary:
    properties:
      average_monthly_income:
        example: 3200
        type: number
      average_monthly_spending:
        example: 2410.55
        type: number
      discretionary_spending:
        example: 2491.65
        type: number
      essential_spending:
        example: 4740
        type: number
      month_count:
        description: |-
          MonthCount is how many distinct months actually carry data in the
          window, which can be fewer than the months the window spans.
        example: 3
        type: integer
      net_savings:
        example: 2368.35
        type: number
      recurring_spending:
        example: 1830
        type: number
      savings_rate:
        description: |-
          SavingsRate is NetSavings over TotalIncome, 0 when there is no
          income in the window.
        example: 0.2467
        type: number
      total_income:
        example: 9600
        type: number
      total_spending:
        example: 7231.65
        type: number
      transaction_count:
        example: 184
        type: integer
    type: object
info:
  contact: {}
  description: |-
    Ingests bank statement PDFs, runs them through an Azure-hosted Mistral OCR endpoint and an Azure OpenAI deployment, and serves the resulting analysis and chart data.

    Errors are returned as RFC 7807 problem documents with the media type application/problem+json.
  title: bankstmt-analyzer API
  version: "1.0"
paths:
...
  /api/v1/uploads/{id}/visualization:
    get:
      description: Returns chart data for a completed upload. The `present` series
        are recomputed from the recorded transactions for the requested window, so
        they always match the window asked for; `forecast` is the model's projection,
        served as stored. The default window is the last 3 months that carry data.
        `months` (1-12) counts back from the last month with data; `from` and `to`
        (YYYY-MM) override it, and supplying just one of them anchors that end.
      parameters:
      - description: Upload id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - default: 3
        description: Number of months to include, 1-12
        in: query
        name: months
        type: integer
      - description: First month, YYYY-MM
        in: query
        name: from
        type: string
      - description: Last month, YYYY-MM
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_http.VisualizationResponse'
        "400":
          description: Invalid id or window parameters
          schema:
            $ref: '#/definitions/internal_http.Problem'
        "404":
          description: No such upload
          schema:
            $ref: '#/definitions/internal_http.Problem'
        "409":
          description: The upload is not completed
          schema:
            $ref: '#/definitions/internal_http.Problem'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http.Problem'
      summary: Get visualization data
      tags:
      - uploads
...
produces:
- application/json
swagger: "2.0"
...